Switch async runtime to smol
async-std is updated to 1.6.2, which adopts smol as the underlying
runtime. The kvmi library now accept streams as smol::Async<T> rather
than arbitrary streams implemeting AsyncRead and AsyncWrite. This is to
simplify the asynchronous invocation of `sendmsg`.

In order to accomodate such an update, the shutdown mechanisms in kvmi,
kvmi-semantic and collector have been reformed. Now to cleanly shutdown,
the users are expected to wait on a JoinHandle. For collector, the job
of cleaning up has been moved out of Drop.

Various other libraries are updated, most notably capnp-rpc and capnp.